requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
configparser==3.7.4
Flask==1.0.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kafka-python==1.4.6
kazoo==2.6.1
MarkupSafe==1.1.1
oauthlib==3.0.1
python-dotenv==0.10.1
requests==2.21.0
requests-oauth==0.4.1
requests-oauthlib==1.2.0
six==1.12.0
tabulate==0.8.3
urllib3==1.24.2
Werkzeug==0.15.2
